% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpRiment.R
\name{ExpR}
\alias{ExpR}
\title{Main class}
\usage{
ExpR(rep.nb, samplers, preprocessors, methods, extractor)
}
\arguments{
\item{samplers}{list of sampler}

\item{preprocessors}{list of preprocessors}

\item{methods}{list of method}

\item{extractor}{the extractor}
}
\value{
an ExpR object
}
\description{
This class define an experiment.
}
\author{
cayek
}
