% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{ExpRmethod}
\alias{ExpRmethod}
\title{Method Constructor}
\usage{
ExpRmethod(...)
}
\arguments{
\item{...}{parameters}
}
\value{
an object of class method
}
\description{
Method Constructor
}
\author{
cayek
}
