% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpRiment.R
\name{ExpRmouline.ExpR}
\alias{ExpRmouline.ExpR}
\title{Run the experiment}
\usage{
\method{ExpRmouline}{ExpR}(expr)
}
\arguments{
\item{expr}{an ExpR class object.}
}
\value{
a tidy tibble with wanted information.
}
\description{
Run the experiment
}
\author{
cayek
}
