% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{ExpRsampler}
\alias{ExpRsampler}
\title{Samler Constructor}
\usage{
ExpRsampler(...)
}
\arguments{
\item{...}{parameters}
}
\value{
an object of class method
}
\description{
Samler Constructor
}
\author{
cayek
}
