% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{+.ExpRlistOperand}
\alias{+.ExpRlistOperand}
\title{Concat list}
\usage{
\method{+}{ExpRlistOperand}(list1, list2)
}
\arguments{
\item{list1}{list 1}

\item{list2}{list 2}
}
\value{
a concatenation of both list
}
\description{
Concat list
}
\examples{

list3 <- list(a = 1) + list(b = 2)

}
\author{
cayek
}
