% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{*.ExpRoperand}
\alias{*.ExpRoperand}
\title{Expand list1 with element of list 2}
\usage{
\method{*}{ExpRoperand}(list1, list2)
}
\arguments{
\item{list1}{list 1}

\item{list2}{list 2}
}
\value{
expanded list
}
\description{
Expand list1 with element of list 2
}
\examples{
class(list1) <- "ExpRoperand"
list2 <- list(b = 2)
class(list2) <- "ExpRoperand"
list1 + list2
}
\author{
cayek
}
